refactor: 로그인 요청 메서드 school-http-client 서비스로 분리
